Agility plugin bug fixes - moved supply state and all supply actions into AgilitySupplyManager
