+++ src/main/java/net/runelite/client/plugins/microbot/agility/AgilityScript.java
@@ -13,18 +13,13 @@ import net.runelite.client.plugins.microbot.agility.courses.GnomeStrongholdCours
 import net.runelite.client.plugins.microbot.agility.courses.PrifddinasCourse;
 import net.runelite.client.plugins.microbot.agility.courses.WerewolfCourse;
 import net.runelite.client.plugins.microbot.agility.enums.AgilityCourse;
-import net.runelite.client.plugins.microbot.agility.enums.AgilityFoodOption;
-import net.runelite.client.plugins.microbot.agility.enums.AgilityPotionOption;
 import net.runelite.client.plugins.microbot.api.tileitem.models.Rs2TileItemModel;
 import net.runelite.client.plugins.microbot.util.antiban.Rs2Antiban;
 import net.runelite.client.plugins.microbot.util.antiban.Rs2AntibanSettings;
-import net.runelite.client.plugins.microbot.util.bank.Rs2Bank;
 import net.runelite.client.plugins.microbot.util.camera.Rs2Camera;
 import net.runelite.client.plugins.microbot.util.gameobject.Rs2GameObject;
 import net.runelite.client.plugins.microbot.util.inventory.Rs2Inventory;
-import net.runelite.client.plugins.microbot.util.inventory.Rs2ItemModel;
 import net.runelite.client.plugins.microbot.util.magic.Rs2Magic;
-import net.runelite.client.plugins.microbot.util.misc.Rs2Food;
 import net.runelite.client.plugins.microbot.util.player.Rs2Player;
 import net.runelite.client.plugins.microbot.util.walker.Rs2Walker;
 
@@ -50,19 +45,10 @@ public class AgilityScript extends Script
 	private AgilityCourseHandler activeHandler = null;
 	private static final int MARK_OF_GRACE_SEARCH_DISTANCE = 30;
 	private static final int MARK_OF_GRACE_PICKUP_TIMEOUT = 5000;
-	private static final long BANK_RETRY_COOLDOWN_MS = 30_000;
-	private static final int SUMMER_PIE_ID = ItemID.SUMMER_PIE;
-	private static final int[] STAMINA_POTION_IDS = {
-		ItemID._4DOSESTAMINA,
-		ItemID._3DOSESTAMINA,
-		ItemID._2DOSESTAMINA,
-		ItemID._1DOSESTAMINA
-	};
 	private WorldPoint pendingMarkOfGraceLocation = null;
 	private int pendingMarkOfGraceCount = 0;
 	private long pendingMarkOfGraceStartedAt = 0;
-	private long lastBankFailureAt = 0;
-	private boolean lastFoodBankUnavailable = false;
+	private final AgilitySupplyManager supplyManager;
 	private volatile boolean shuttingDown = false;
 
 	@Inject
@@ -70,6 +56,7 @@ public class AgilityScript extends Script
 	{
 		this.plugin = plugin;
 		this.config = config;
+		this.supplyManager = new AgilitySupplyManager(plugin, config, this::isShuttingDown, this::shutdown);
 	}
 
 	@Override
@@ -85,8 +72,7 @@ public class AgilityScript extends Script
 		activeHandler = null;
 		startPoint = null;
 		initialPlayerLocation = null;
-		lastBankFailureAt = 0;
-		lastFoodBankUnavailable = false;
+		supplyManager.reset();
 		clearPendingMarkOfGrace();
 
 		if (mainScheduledFuture != null && !mainScheduledFuture.isDone())
@@ -152,15 +138,28 @@ public class AgilityScript extends Script
 					return;
 				}
 
-				if (handleSummerPies(courseHandler))
+				if (supplyManager.handleSummerPies(courseHandler))
 				{
 					Microbot.log("Early return: Handling summer pies");
 					return;
 				}
 
-				if (!plugin.hasRequiredLevel() && handleBanking(courseHandler, playerWorldLocation))
+				if (supplyManager.handleFoodOrHealthSafety())
+				{
+					Microbot.log("Early return: Handling agility safety");
+					return;
+				}
+
+				boolean hasRequiredLevel = plugin.hasRequiredLevel();
+				if (supplyManager.handlePreLevelCheck(courseHandler, playerWorldLocation, hasRequiredLevel))
 				{
 					Microbot.log("Early return: Banking agility supplies");
+					return;
+				}
+
+				if (supplyManager.handleBeforeObstacle(courseHandler, playerWorldLocation))
+				{
+					Microbot.log("Early return: Handling agility supplies");
 					return;
 				}
 
@@ -171,7 +170,7 @@ public class AgilityScript extends Script
 					shutdown();
 					return;
 				}
-				
+
 				// Check coin requirement for BrimhavenSpike course (only before payment)
 				if (courseHandler instanceof BrimhavenSpikeCourse)
 				{
@@ -191,21 +190,6 @@ public class AgilityScript extends Script
 				}
 				final int currentAgilityXp = Microbot.getClient().getSkillExperience(Skill.AGILITY);
 
-				if (handleFood())
-				{
-					Microbot.log("Early return: Handling food");
-					return;
-				}
-				if (handleBanking(courseHandler, playerWorldLocation))
-				{
-					Microbot.log("Early return: Banking agility supplies");
-					return;
-				}
-				if (handleHealthSafety())
-				{
-					Microbot.log("Early return: Waiting for safe hitpoints");
-					return;
-				}
 				if (lootMarksOfGrace(courseHandler))
 				{
 					Microbot.log("Early return: Looting marks of grace");
@@ -406,8 +390,7 @@ public class AgilityScript extends Script
 			activeHandler.reset();
 			startPoint = activeHandler.getStartPoint();
 			lastAgilityXp = Microbot.getClient().getSkillExperience(Skill.AGILITY);
-			lastBankFailureAt = 0;
-			lastFoodBankUnavailable = false;
+			supplyManager.reset();
 		}
 		return activeHandler;
 	}
@@ -522,373 +505,6 @@ public class AgilityScript extends Script
 			.where(Rs2TileItemModel::isLootAble)
 			.where(item -> markLocation.equals(item.getWorldLocation()))
 			.first() != null;
-	}
-
-	private boolean handleFood()
-	{
-		if (config.hitpoints() <= 0)
-		{
-			return false;
-		}
-
-		if (Rs2Player.getHealthPercentage() > config.hitpoints())
-		{
-			return false;
-		}
-
-		List<Rs2ItemModel> foodItems = plugin.getInventoryFood();
-		if (foodItems.isEmpty())
-		{
-			if (needsFoodBanking() || config.hpSafetyWait())
-			{
-				return false;
-			}
-			Microbot.showMessage("Hitpoints are below the configured threshold and no food was found. Stopping agility.");
-			shutdown();
-			return true;
-		}
-		Rs2ItemModel foodItem = foodItems.get(0);
-
-		Rs2Inventory.interact(foodItem, foodItem.getName().toLowerCase().contains("jug of wine") ? "drink" : "eat");
-		Rs2Inventory.waitForInventoryChanges(1800);
-
-		if (Rs2Inventory.contains(ItemID.JUG_EMPTY))
-		{
-			Rs2Inventory.dropAll(ItemID.JUG_EMPTY);
-		}
-		return true;
-	}
-
-	private boolean handleHealthSafety()
-	{
-		if (!shouldWaitForHealthInsteadOfBanking())
-		{
-			return false;
-		}
-
-		int resumeAt = Math.max(config.hpSafetyResumeAt(), config.hpSafetyWaitBelow());
-		Microbot.status = "Waiting for hitpoints";
-		sleepUntil(() -> shuttingDown
-			|| !plugin.getInventoryFood().isEmpty()
-			|| Rs2Player.getHealthPercentage() >= resumeAt, 10_000);
-		return true;
-	}
-
-	private boolean handleBanking(AgilityCourseHandler courseHandler, WorldPoint playerWorldLocation)
-	{
-		if (!shouldBankForSupplies(courseHandler))
-		{
-			return false;
-		}
-		if (shouldWaitForHealthInsteadOfBanking())
-		{
-			return false;
-		}
-		if (Rs2Player.isMoving() || Rs2Player.isAnimating())
-		{
-			return true;
-		}
-		if (!canAttemptBanking(courseHandler, playerWorldLocation))
-		{
-			Microbot.status = "Finishing course before banking";
-			return false;
-		}
-
-		long now = System.currentTimeMillis();
-		if (lastBankFailureAt > 0 && now - lastBankFailureAt < BANK_RETRY_COOLDOWN_MS)
-		{
-			Microbot.status = "Waiting before retrying bank";
-			return true;
-		}
-
-		Microbot.status = "Banking agility supplies";
-		lastFoodBankUnavailable = false;
-		if (!Rs2Bank.walkToBankAndUseBank())
-		{
-			lastBankFailureAt = now;
-			Microbot.log("Unable to reach or open bank for agility supplies. Retrying after cooldown.");
-			return true;
-		}
-
-		boolean success = withdrawConfiguredSupplies(courseHandler);
-		Rs2Bank.closeBank();
-		lastBankFailureAt = success ? 0 : System.currentTimeMillis();
-		return true;
-	}
-
-	private boolean withdrawConfiguredSupplies(AgilityCourseHandler courseHandler)
-	{
-		boolean success = true;
-		if (needsFoodBanking())
-		{
-			success &= withdrawConfiguredFood();
-		}
-		if (needsPotionBanking())
-		{
-			success &= withdrawConfiguredPotion();
-		}
-		if (needsSummerPieBanking(courseHandler))
-		{
-			success &= withdrawSummerPies();
-		}
-		return success;
-	}
-
-	private boolean withdrawConfiguredFood()
-	{
-		int targetAmount = config.foodWithdrawAmount();
-		int currentFood = plugin.getInventoryFood().size();
-		int missingFood = targetAmount - currentFood;
-		if (missingFood <= 0)
-		{
-			return true;
-		}
-		if (Rs2Inventory.emptySlotCount() <= 0)
-		{
-			Microbot.showMessage("Inventory is full. Unable to withdraw agility food.");
-			return false;
-		}
-
-		int amountToWithdraw = Math.min(missingFood, Rs2Inventory.emptySlotCount());
-
-		AgilityFoodOption option = config.bankFood();
-		if (option.isAuto())
-		{
-			return withdrawAutoFood(amountToWithdraw, currentFood);
-		}
-
-		Rs2Food food = option.getFood();
-		if (food == null)
-		{
-			return false;
-		}
-		if (!Rs2Bank.hasBankItem(food.getId(), amountToWithdraw))
-		{
-			lastFoodBankUnavailable = true;
-			Microbot.showMessage("No " + food.getName() + " found in the bank.");
-			return false;
-		}
-
-		if (!Rs2Bank.withdrawX(food.getId(), amountToWithdraw))
-		{
-			return false;
-		}
-		return sleepUntil(() -> plugin.getInventoryFood().size() >= currentFood + amountToWithdraw, 2400);
-	}
-
-	private boolean withdrawAutoFood(int amountToWithdraw, int currentFood)
-	{
-		int remaining = amountToWithdraw;
-		int expectedFood = currentFood;
-
-		List<AgilityFoodOption> foodOptions = Arrays.stream(AgilityFoodOption.values())
-			.filter(foodOption -> !foodOption.isNone() && !foodOption.isAuto())
-			.filter(foodOption -> foodOption.getFood() != null)
-			.sorted(Comparator.comparingInt((AgilityFoodOption foodOption) -> foodOption.getFood().getHeal()).reversed())
-			.collect(Collectors.toList());
-
-		for (AgilityFoodOption option : foodOptions)
-		{
-			if (remaining <= 0 || Rs2Inventory.emptySlotCount() <= 0)
-			{
-				break;
-			}
-
-			Rs2Food food = option.getFood();
-			int available = getBankItemQuantity(food.getId());
-			if (available <= 0)
-			{
-				continue;
-			}
-
-			int withdrawAmount = Math.min(Math.min(remaining, available), Rs2Inventory.emptySlotCount());
-			if (!Rs2Bank.withdrawX(food.getId(), withdrawAmount))
-			{
-				return false;
-			}
-
-			int expectedAfterWithdraw = expectedFood + withdrawAmount;
-			if (!sleepUntil(() -> plugin.getInventoryFood().size() >= expectedAfterWithdraw, 2400))
-			{
-				return false;
-			}
-			expectedFood = expectedAfterWithdraw;
-			remaining = currentFood + amountToWithdraw - plugin.getInventoryFood().size();
-		}
-
-		if (remaining > 0)
-		{
-			lastFoodBankUnavailable = true;
-			Microbot.showMessage("Not enough configured food was found in the bank.");
-			return false;
-		}
-		return true;
-	}
-
-	private boolean withdrawConfiguredPotion()
-	{
-		AgilityPotionOption option = config.bankPotion();
-		int targetAmount = config.potionWithdrawAmount();
-		int currentPotions = getStaminaPotionCount();
-		int missingPotions = targetAmount - currentPotions;
-		if (missingPotions <= 0)
-		{
-			return true;
-		}
-		if (Rs2Inventory.emptySlotCount() <= 0)
-		{
-			Microbot.showMessage("Inventory is full. Unable to withdraw agility potions.");
-			return false;
-		}
-
-		int amountToWithdraw = Math.min(missingPotions, Rs2Inventory.emptySlotCount());
-		if (option.isNone() || option.getItemId() <= 0)
-		{
-			return false;
-		}
-		if (!Rs2Bank.hasBankItem(option.getItemId(), amountToWithdraw))
-		{
-			Microbot.showMessage("No " + option + " found in the bank.");
-			return false;
-		}
-
-		if (!Rs2Bank.withdrawX(option.getItemId(), amountToWithdraw))
-		{
-			return false;
-		}
-		return sleepUntil(() -> getStaminaPotionCount() >= currentPotions + amountToWithdraw, 2400);
-	}
-
-	private boolean withdrawSummerPies()
-	{
-		int targetAmount = config.summerPieWithdrawAmount();
-		int currentPies = plugin.getSummerPies().size();
-		int missingPies = targetAmount - currentPies;
-		if (missingPies <= 0)
-		{
-			return true;
-		}
-		if (Rs2Inventory.emptySlotCount() <= 0)
-		{
-			Microbot.showMessage("Inventory is full. Unable to withdraw summer pies.");
-			return false;
-		}
-
-		int amountToWithdraw = Math.min(missingPies, Rs2Inventory.emptySlotCount());
-		if (!Rs2Bank.hasBankItem(SUMMER_PIE_ID, amountToWithdraw))
-		{
-			Microbot.showMessage("No summer pies found in the bank.");
-			return false;
-		}
-		if (!Rs2Bank.withdrawX(SUMMER_PIE_ID, amountToWithdraw))
-		{
-			return false;
-		}
-		return sleepUntil(() -> plugin.getSummerPies().size() >= currentPies + amountToWithdraw, 2400);
-	}
-
-	private boolean shouldBankForSupplies(AgilityCourseHandler courseHandler)
-	{
-		return needsFoodBanking() || needsPotionBanking() || needsSummerPieBanking(courseHandler);
-	}
-
-	private boolean needsFoodBanking()
-	{
-		AgilityFoodOption food = config.bankFood();
-		return isFoodBankingConfigured()
-			&& plugin.getInventoryFood().size() <= config.bankFoodAt()
-			&& plugin.getInventoryFood().size() < config.foodWithdrawAmount();
-	}
-
-	private boolean isFoodBankingConfigured()
-	{
-		AgilityFoodOption food = config.bankFood();
-		return !food.isNone()
-			&& config.foodWithdrawAmount() > 0
-			&& config.bankFoodAt() > 0;
-	}
-
-	private boolean needsPotionBanking()
-	{
-		AgilityPotionOption potion = config.bankPotion();
-		return !potion.isNone()
-			&& config.potionWithdrawAmount() > 0
-			&& getStaminaPotionCount() < config.potionWithdrawAmount();
-	}
-
-	private boolean needsSummerPieBanking(AgilityCourseHandler courseHandler)
-	{
-		return courseHandler != null
-			&& courseHandler.canBeBoosted()
-			&& config.summerPieWithdrawAmount() > 0
-			&& plugin.getSummerPies().size() < config.summerPieWithdrawAmount();
-	}
-
-	private int getStaminaPotionCount()
-	{
-		return Rs2Inventory.count(item -> item != null && Arrays.stream(STAMINA_POTION_IDS).anyMatch(id -> id == item.getId()));
-	}
-
-	private boolean shouldWaitForHealthInsteadOfBanking()
-	{
-		return config.hpSafetyWait()
-			&& plugin.getInventoryFood().isEmpty()
-			&& Rs2Player.getHealthPercentage() < config.hpSafetyWaitBelow()
-			&& (!isFoodBankingConfigured() || lastFoodBankUnavailable);
-	}
-
-	private boolean canAttemptBanking(AgilityCourseHandler courseHandler, WorldPoint playerWorldLocation)
-	{
-		if (!config.bankOnlyAtCourseStart())
-		{
-			return true;
-		}
-		if (courseHandler == null || playerWorldLocation == null || courseHandler.getStartPoint() == null)
-		{
-			return false;
-		}
-		return courseHandler.getClientPlane() == 0
-			&& courseHandler.getCurrentObstacleIndex() == 0
-			&& playerWorldLocation.distanceTo(courseHandler.getStartPoint()) <= 12;
-	}
-
-	private int getBankItemQuantity(int itemId)
-	{
-		return Rs2Bank.bankItems().stream()
-			.filter(item -> item != null && item.getId() == itemId)
-			.mapToInt(Rs2ItemModel::getQuantity)
-			.sum();
-	}
-
-	private boolean handleSummerPies(AgilityCourseHandler courseHandler)
-	{
-		if (!courseHandler.canBeBoosted())
-		{
-			return false;
-		}
-		if (courseHandler.getCurrentObstacleIndex() > 0)
-		{
-			return false;
-		}
-		if (Rs2Player.getBoostedSkillLevel(Skill.AGILITY) >= courseHandler.getRequiredLevel())
-		{
-			return false;
-		}
-
-		List<Rs2ItemModel> summerPies = plugin.getSummerPies();
-		if (summerPies.isEmpty())
-		{
-			return false;
-		}
-		Rs2ItemModel summerPie = summerPies.get(0);
-
-		Rs2Inventory.interact(summerPie, "eat");
-		Rs2Inventory.waitForInventoryChanges(1800);
-		if (Rs2Inventory.contains(ItemID.PIEDISH))
-		{
-			Rs2Inventory.dropAll(ItemID.PIEDISH);
-		}
-		return true;
 	}
 
 	private boolean shouldPerformAlch()
