+++ src/main/java/net/runelite/client/plugins/microbot/agility/MicroAgilityPlugin.java
@@ -34,7 +34,7 @@ import java.util.stream.Collectors;
 @Slf4j
 public class MicroAgilityPlugin extends Plugin
 {
-	public static final String version = "1.3.0";
+	public static final String version = "1.3.2";
 	@Inject
 	private MicroAgilityConfig config;
 	@Inject
